+++ backend/src/lib.rs
@@ -34,7 +34,6 @@ pub mod inspect;
 pub mod install;
 pub mod logs;
 pub mod manager;
-pub mod marketplace;
 pub mod middleware;
 pub mod migration;
 pub mod net;
@@ -43,7 +42,7 @@ pub mod os_install;
 pub mod prelude;
 pub mod procedure;
 pub mod properties;
-pub mod registryadmin;
+pub mod registry;
 pub mod s9pk;
 pub mod setup;
 pub mod shutdown;
@@ -80,7 +79,7 @@ pub fn echo(#[arg] message: String) -> Result<String, RpcError> {
     disk::disk,
     notifications::notification,
     backup::backup,
-    marketplace::marketplace,
+    registry::marketplace::marketplace,
 ))]
 pub fn main_api() -> Result<(), RpcError> {
     Ok(())
@@ -126,7 +125,7 @@ pub fn package() -> Result<(), RpcError> {
     developer::verify,
     developer::init,
     inspect::inspect,
-    registryadmin::publish,
+    registry::admin::publish,
 ))]
 pub fn portable_api() -> Result<(), RpcError> {
     Ok(())
