+++ backend/src/registry/admin.rs
